+++ packages/flutter/lib/plass_ui.dart
@@ -58,6 +58,7 @@ export 'src/components/combobox/pl_combobox.dart';
 export 'src/components/command_palette/pl_command_palette.dart';
 export 'src/components/confirm/pl_confirm.dart';
 export 'src/components/container/pl_container.dart';
+export 'src/components/data_list/pl_data_list.dart';
 export 'src/components/date_picker/pl_date_picker.dart';
 export 'src/components/date_range_picker/pl_date_range_picker.dart';
 export 'src/components/date_time_picker/pl_date_time_picker.dart';